updated the location for CP1CP2 image
AP part is not specified in the handout. Suggest to delete the image. And maybe explain shortly()

The original AP fails when the attacker performs a replay attack. The malicious server can trick the honest client into believing that he/she has succesfully uploaded their data onto the server when the server has in fact not received anything. 

To resolve this issue, in ClientCP1 and ClientCP2, the client sends a freshly generated nonce (SecureRandom BigInteger) to the server. The server encrypts this nonce with its private key and sends it back to
the client. Since the nonce is different for every session, the malicious server cannot save the transmitted information to trick the client in future sessions.